Migration to VS2010, - Multithreaded rendering is officialy bugged, don't use it!!! + Few editions to renderer like start/end time to render from log (However, log file still will be parsed from beggining (No fix for this ATM)) + New parameter to change font size for nick (usefull for HD resolution video rendering, so it would downscale better..) + Possible fix for corrupted logs (further testing required, it might turn out, that after log corruption seek to the next INIT might be needed.)

+++ trunk/SocialGraphDll/SocialGraphDll/GraphVideo.h
@@ -9,7 +9,7 @@ public:
 	~GraphVideo();
 	void addEdge(const std::string *ln1, const std::string *ln2, double weight,int activity);
 	void renderVideo();
-	void renderFrames(int &nextRender,int timestamp);
+	void renderFrames(double &nextRender,int timestamp);
 	void renderRelocateNode(Node *n);
 	void drawImage(std::wstring *fWPath,int szClock);
 	void makeImage(int iterations, std::wstring *output,int tNow = (int)time(NULL));
@@ -19,6 +19,6 @@ public:
 	void decay(double d, int tNow);
 private:
 	int vidRendFrame;
-	int vidSecsPerFrame;
+	double vidSecsPerFrame;
 	GraphRendererQueue *grq;
 };
